requirements.txt
APScheduler==3.6.3
certifi==2023.5.7
cffi==1.14.3
cryptography==40.0.2
decorator==4.4.2
pycparser==2.20
python-telegram-bot==13.9
pytz==2020.4
six==1.15.0
tornado==6.3.2
tzlocal==2.1
pyyaml